fixing stuff in the virtual PoMos

+++ src/core/datatypes/phylogenetics/ratematrix/RateMatrix_PoMo2N.cpp
@@ -329,7 +329,7 @@ std::vector<double> RateMatrix_PoMo2N::getStationaryFrequencies( void ) const
 }
 
 
-void RateMatrix_PoMo2N::setNeff( long ni )
+void RateMatrix_PoMo2N::setNeff( double ni )
 {
     N_eff = ni;
